reuse setup_logging and move it to basic utils
