+++ utils/utils.py
@@ -5,9 +5,6 @@ import urllib.parse
 import calendar
 import datetime
 import sys
-import getopt
-import calendar
-import logging
 
 from PyPDF2 import PdfFileReader
 from xml.parsers.expat import ExpatError
@@ -343,53 +340,6 @@ def extract_links_from_pdf(fileobj):
                 if href:
                     links.append(href)
     return links
-
-def setup_logging(loglevel, logfile):
-    leveldict = {'critical': logging.CRITICAL, 'error': logging.ERROR, \
-                 'warning': logging.WARNING,   'info': logging.INFO, \
-                 'debug': logging.DEBUG}
-
-    logfmt  = '%(asctime)s: %(name)s: %(levelname)s %(message)s'
-    datefmt = '%Y-%m-%d %H:%M:%S'
-    if logfile:
-        logging.basicConfig(\
-            level   = leveldict[loglevel], \
-            format  = logfmt, \
-            filename = logfile, \
-            datefmt = datefmt \
-        )
-    else:
-        logging.basicConfig(\
-            level   = leveldict[loglevel], \
-            format  = logfmt, \
-            datefmt = datefmt \
-        )
-
-
-# multiprocessing module on macs does forking weirdly, so logging doesn't work as expected
-# if on darwin call this function inside the task to reinitialize logging
-def setup_logging_if_needed():
-    platform = sys.platform
-    if platform != 'darwin':
-        return
-    optlist, remlist = getopt.getopt(sys.argv[1:], 'ad:D:l:mnf:p:t:T:hrs:W:')
-    datadir    = None
-    debuglevel = 'info'
-    filename = None
-    for o, v in optlist:
-        if o == '-D':
-            datadir = v
-        elif o == '-l':
-            debuglevel = v
-        elif o == '-f':
-            filename = v
-    statsdir = os.path.join(datadir, 'stats')
-    mk_dir(statsdir)
-
-    if filename:
-        filename = os.path.join(statsdir, filename)
-    setup_logging(debuglevel, filename)
-
